[components][finsh] Decouple RT_USING_DEVICE dependency for msh feature
