+++ components/finsh/symbol.c
@@ -10,7 +10,7 @@
  
 #include <rtthread.h>
 
-#ifdef RT_USING_FINSH
+#if defined(RT_USING_FINSH) && !defined(FINSH_USING_SYMTAB)
 
 #include "finsh.h"
 
@@ -27,12 +27,6 @@ long list_msgqueue(void);
 long list_mempool(void);
 long list_timer(void);
 
-#ifdef FINSH_USING_SYMTAB
-struct finsh_syscall *_syscall_table_begin  = NULL;
-struct finsh_syscall *_syscall_table_end    = NULL;
-struct finsh_sysvar *_sysvar_table_begin    = NULL;
-struct finsh_sysvar *_sysvar_table_end      = NULL;
-#else
 struct finsh_syscall _syscall_table[] =
 {
     {"hello", hello},
@@ -67,7 +61,6 @@ struct finsh_syscall *_syscall_table_end   = &_syscall_table[sizeof(_syscall_tab
 
 struct finsh_sysvar *_sysvar_table_begin  = NULL;
 struct finsh_sysvar *_sysvar_table_end    = NULL;
-#endif
 
-#endif /* RT_USING_FINSH */
+#endif /* RT_USING_FINSH && !FINSH_USING_SYMTAB */
