+++ components/utilities/ulog/backend/console_be.c
@@ -22,9 +22,9 @@ static struct ulog_backend console;
 void ulog_console_backend_output(struct ulog_backend *backend, rt_uint32_t level, const char *tag, rt_bool_t is_raw,
         const char *log, size_t len)
 {
+#ifdef RT_USING_DEVICE
     rt_device_t dev = rt_console_get_device();
 
-#ifdef RT_USING_DEVICE
     if (dev == RT_NULL)
     {
         rt_hw_console_output(log);
